Keep channel tuning fractional, and fix the drum and key-shift layers
RPN fine tuning resolves to 100/8192 of a cent, and the port was throwing
almost all of that away.  ss_channel_set_tuning rounded its argument to whole
cents, and channel_tuning_cents was an int that truncated the sum of the tuning
custom controllers again; render_voice then narrowed what was left into a float
accumulator.  Upstream holds every one of those as a number.  All three are
widened, and the two roundings are gone.

Two neighbouring bugs in the same computation.  A drum channel takes the
channel *system* fine tune and key shift alone -- the port was adding the
channel MIDI layer to it as well.  And the key shift is truncated toward zero,
as upstream's Math.trunc does, not floored, which sent a fractional negative
shift a semitone further down than intended.

rpn_fine_tuning_test 37.50 -> 118.44 dB.  Nothing else moves, the three C tests
pass, and near-identical rises to seventeen of thirty-one.

rpn_tuning_real-time_test is untouched at 6.28 dB and is a different bug: RPN
coarse tuning should set the channel's key shift, which reselects the sample
zone, where the port applies it as a semitone tuning offset.  The test says so
itself -- "should be treated as non-realtime (key shift)".

+++ spessasynth_core/src/synthesizer/audio_engine/channel/data_entry.c
@@ -20,9 +20,9 @@
 extern void ss_channel_set_generator_offset(SS_MIDIChannel *ch, SS_GeneratorType gen, int val, double time);
 
 extern void ss_channel_reset_parameters_to_defaults(SS_MIDIChannel *ch);
-extern void ss_channel_set_custom_controller(SS_MIDIChannel *ch, SS_CustomController type, float val);
-extern void ss_channel_set_tuning(SS_MIDIChannel *ch, float cents);
-extern void ss_channel_set_modulation_depth(SS_MIDIChannel *ch, float cents);
+extern void ss_channel_set_custom_controller(SS_MIDIChannel *ch, SS_CustomController type, double val);
+extern void ss_channel_set_tuning(SS_MIDIChannel *ch, double cents);
+extern void ss_channel_set_modulation_depth(SS_MIDIChannel *ch, double cents);
 extern void ss_channel_set_pitch_wheel_range(SS_MIDIChannel *ch, int value);
 extern void ss_channel_nrpn_awe32(SS_MIDIChannel *ch, int param_lsb, int data_value, double time);
 
@@ -53,14 +53,14 @@ void ss_channel_data_entry(SS_MIDIChannel *ch, double time) {
 				/* Fine-tuning */
 				const int final_tuning = data_value - 8192;
 				/* Resolution is 100/8192 cents */
-				ss_channel_set_tuning(ch, (float)final_tuning / 81.92f);
+				ss_channel_set_tuning(ch, (double)final_tuning / 81.92);
 				break;
 			}
 
 				/* Modulation depth */
 			case SS_RPN_MODULATION_DEPTH:
 				/* Cents, so data / 128 * 100 is data / 1.28 */
-				ss_channel_set_modulation_depth(ch, (float)data_value / 1.28f);
+				ss_channel_set_modulation_depth(ch, (double)data_value / 1.28);
 				break;
 
 			case SS_RPN_RESET_PARAMETERS:
